+++ RELEASE_NOTES.md
@@ -93,6 +93,7 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `KeyboardAction` `.backspace` now triggers on `.press` instead of `.release`.
 * `KeyboardLayoutConfiguration` now behaves the same on all platforms.
 * `KeyboardSettingsLink` has ben rewritten to only use plain SwiftUI code that works with extensions as well.
+* `StandardKeyboardAppearance` now applies medium font weights to the Gregorian ABC keyboard. 
 
 ### 💥 Breaking changes